# api/v1/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for DixitMe - Online Dixit Card Game\nThis API provides endpoints for managing players, games, and real-time gameplay through WebSocket connections.",
        "title": "DixitMe API",
        "contact": {
            "name": "DixitMe API Support",
            "email": "support@dixitme.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/games/add-bot": {
            "post": {
                "description": "Add an AI bot player to an existing game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Add bot to game",
                "parameters": [
                    {
                        "description": "Bot information",
                        "name": "bot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddBotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/games/leave": {
            "post": {
                "description": "Allow a player to leave a game they are currently in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Leave game",
                "parameters": [
                    {
                        "description": "Leave game information",
                        "name": "leave",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LeaveGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/games/remove-player": {
            "delete": {
                "description": "Remove a player or bot from an existing game (only in waiting status)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Remove player from game",
                "parameters": [
                    {
                        "description": "Player removal information",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RemovePlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/games/{room_code}": {
            "delete": {
                "description": "Delete an entire game (only allowed by lobby manager/creator)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Delete game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room code",
                        "name": "room_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/bots/stats": {
            "get": {
                "description": "Get statistics about AI bot performance and usage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Get bot statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BotStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/games": {
            "get": {
                "description": "Get a list of all games with their current status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get all games",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetGamesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/games/{room_code}": {
            "get": {
                "description": "Get game information and check if it's currently live",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game by room code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game room code",
                        "name": "room_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetGameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.AddBotRequest": {
            "type": "object",
            "required": [
                "room_code"
            ],
            "properties": {
                "bot_level": {
                    "description": "easy, medium, hard",
                    "type": "string"
                },
                "room_code": {
                    "type": "string"
                }
            }
        },
        "handlers.BotStatsResponse": {
            "type": "object",
            "properties": {
                "active_bots": {
                    "type": "integer"
                },
                "bot_performance": {
                    "type": "object",
                    "additionalProperties": true
                },
                "bots_by_level": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "total_bots": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.GetGameResponse": {
            "type": "object",
            "properties": {
                "game": {
                    "$ref": "#/definitions/models.Game"
                },
                "is_live": {
                    "type": "boolean"
                }
            }
        },
        "handlers.GetGamesResponse": {
            "type": "object",
            "properties": {
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Game"
                    }
                }
            }
        },
        "handlers.LeaveGameRequest": {
            "type": "object",
            "required": [
                "room_code"
            ],
            "properties": {
                "room_code": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RemovePlayerRequest": {
            "type": "object",
            "required": [
                "player_id",
                "room_code"
            ],
            "properties": {
                "player_id": {
                    "type": "string"
                },
                "room_code": {
                    "type": "string"
                }
            }
        },
...
        "models.AuthType": {
            "type": "string",
            "enum": [
                "guest",
                "password",
                "google"
            ],
            "x-enum-varnames": [
                "AuthTypeGuest",
                "AuthTypePassword",
                "AuthTypeGoogle"
            ]
        },
...
        "models.CardSubmission": {
            "type": "object",
            "properties": {
                "card_id": {
                    "description": "ID of the card from the deck",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                },
                "round": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GameRound"
                        }
                    ]
                },
                "round_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Game": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "max_rounds": {
                    "description": "3 players * 2 rounds each",
                    "type": "integer"
                },
                "players": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GamePlayer"
                    }
                },
                "room_code": {
                    "type": "string"
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GameRound"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.GameStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.GamePlayer": {
            "type": "object",
            "properties": {
                "game": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Game"
                        }
                    ]
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "player_id": {
                    "type": "string"
                },
                "position": {
                    "description": "Turn order",
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "models.GameRound": {
            "type": "object",
            "properties": {
                "clue": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "game": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Game"
                        }
                    ]
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "round_number": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                },
                "storyteller": {
                    "$ref": "#/definitions/models.Player"
                },
                "storyteller_card": {
                    "description": "Card ID chosen by storyteller",
                    "type": "integer"
                },
                "storyteller_id": {
                    "type": "string"
                },
                "submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CardSubmission"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vote"
                    }
                }
            }
        },
        "models.GameStatus": {
            "type": "string",
            "enum": [
                "waiting",
                "in_progress",
                "completed",
                "abandoned"
            ],
            "x-enum-varnames": [
                "GameStatusWaiting",
                "GameStatusInProgress",
                "GameStatusCompleted",
                "GameStatusAbandoned"
            ]
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "$ref": "#/definitions/models.AuthType"
                },
                "bot_level": {
                    "description": "easy, medium, hard",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.PlayerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "NULL for guest players",
                    "type": "string"
                }
            }
        },
        "models.PlayerType": {
            "type": "string",
            "enum": [
                "human",
                "bot"
            ],
            "x-enum-varnames": [
                "PlayerTypeHuman",
                "PlayerTypeBot"
            ]
        },
        "models.RoundStatus": {
            "type": "string",
            "enum": [
                "storytelling",
                "submitting",
                "voting",
                "scoring",
                "completed"
            ],
            "x-enum-varnames": [
                "RoundStatusStorytelling",
                "RoundStatusSubmitting",
                "RoundStatusVoting",
                "RoundStatusScoring",
                "RoundStatusCompleted"
            ]
        },
...
    }
}